we can now draw a fretboard
I am not /super/ happy with the implementation of this fretboard
but I think that it is a good start. There is likely a way to improve
the way I am going about drawing all of these rectangles, but now
that I have a basic foundation I will move on to drawing notes and
worry about optimizing things after.