capping out selectable chords at 5 notes
I wanted to inhibit the amount of notes we could have in a chord
arbitrarily for now so that I can work on other more basic features.
Currently you should now be able to select up to 5 notes for a chord
and it will highlight each of those notes and then also highlight
any secondary notes with a slightly darker tone of the same color.