changing the rect used for clicks on notes
There was an issue where the rect for the mouse clicks on a note
and the rect for hovering over a note were different. This is because
the lettering for the notes is actually a smaller rect than that of
the rect for the circle. I updated the call to use the circle for the
rect and now when you click it consistently will toggle the note